Custom terms off switch

+++ src/web.api/jobs/delete-drivers-custom-tarrif-job.mjs
@@ -9,6 +9,11 @@ const job = CronJob.from({
     cronTime,
     timeZone,
     onTick: async () => {
+
+        if (process.env.CUSTOM_TERMS == 'OFF') {
+            return
+        }
+
         try {
             await deleteDriversCustomTariff();
         } catch (error) {
